requirements.txt
alembic==1.12.0
factory-boy==3.3.0
fastapi==0.110.0
fastapi-pagination==0.12.17
httpx==0.27.0
psycopg2-binary==2.9.9
pytest==7.4.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
sqladmin==0.16.0
uvicorn==0.27.1